requirements.txt
librosa==0.10.2.post1
mistralai==0.4.1
pillow==10.3.0
pytesseract==0.3.10
streamlit==1.36.0
torch==2.3.1
transformers==4.42.3
